requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.3
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
twilio==8.2.0
celery==5.3.1
redis==4.5.5
phonenumbers==8.13.14
